изменил порт подключения к postgres через Docker с 5432:5432 на 5433:5432; изменил тип latitude и longitude с Double на BigDecimal; изменил версию postgres с 16 на 14.6, flyway с 10.18.0 на 9.22.3, hibernate с 6.3.1.Final на 6.1.5.Final; изменил порядок создания бинов: было Hibernate -> Flyway, стало flyway -> Hibernate;

+++ src/main/java/com/weather/model/dto/LocationDto.java
@@ -1,9 +1,11 @@
 package com.weather.model.dto;
 
+import java.math.BigDecimal;
+
 public record LocationDto(
         String name,
-        Double lat,
-        Double lon,
+        BigDecimal lat,
+        BigDecimal lon,
         String country,
         String state) {
 }
